+++ fast_sync_dart_client/fast_sync_client/example/lib/item/pages/item_provider.dart
@@ -21,6 +21,10 @@ class ItemProvider with ChangeNotifier {
     await syncManager.pull();
   }
 
+  Future<void> sync() async {
+    await syncManager.sync();
+  }
+
   Future<void> push() async {
     await syncManager.push();
   }
@@ -42,9 +46,9 @@ class ItemProvider with ChangeNotifier {
     return item;
   }
 
-  Future<void> getCount() async {
+  Future<int> getCount() async {
     int count = await repository.count();
-    int test = 1 + 1;
+    return count;
   }
 
   Future<void> resetItemRepo() async {