+++ fast_sync_dart_client/fast_sync_client/lib/src/absraction/service/isync_manager.dart
@@ -3,5 +3,6 @@ import 'package:fast_sync_client/src/absraction/models/sync_payload.dart';
 abstract class ISyncManager {
   Future<SyncPayload> push();
   Future<SyncPayload> pull();
+  Future<SyncPayload> sync();
   Future<SyncPayload> hardReset({List<Type>? types});
 }