+++ fast_sync_dart_client/fast_sync_hive_dao/lib/src/data/syncable_object_data_source.dart
@@ -77,7 +77,7 @@ class SyncalbeObjectDataSource<T extends IWithId>
     final Box<T> box = await boxInstance;
     int count = 0;
     try {
-      count = await box.toMap().values.length;
+      count = box.toMap().values.length;
     } catch (err) {
       count = 0;
     }