+++ fast_sync_dart_client/fast_sync_client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## [1.0.1]
+
+- Fix setTypeSyncZone for global syncZone
+- Fix lint errors
+- remove uneeded package 'getIt'
+
 ## [1.0.0]
 
 - initial release